Added Formula for Calculating Initial Velocity
Added a formula for calculating initial velocity given the horizontal and vertical position of a trajectory (which will be the hole of the goal), the launch angle, and the acceleration of gravity. The initial velocity will be converted into the rpm of the flywheel and then into flywheel motor percent, which means we can now calculate the speed at which we need to spin our flywheel motors to get a ball into the goal anywhere on the field.